# API/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Authentication API",
        "version": "1.0.0",
        "description": "API for user authentication and registration"
    },
    "paths": {
        "/register": {
            "post": {
                "summary": "Register a new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securePassword123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+48123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "ul. Przyk\u0142adowa 123, Warszawa"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password",
                                    "phone_number",
                                    "address",
                                    "date_of_birth"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "User registered successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User with this email already exists",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "User with this email already exists"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login a user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securePassword123"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Login successful"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Invalid credentials"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Get list of users",
                "responses": {
                    "200": {
                        "description": "List of users returned successfully",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "email": "user1@example.com",
                                        "phone_number": "+48123456789",
                                        "address": "ul. Przyk\u0142adowa 123, Warszawa",
                                        "date_of_birth": "1990-01-15"
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the user to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details returned successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "email": "user1@example.com",
                                    "phone_number": "+48123456789",
                                    "address": "ul. Przyk\u0142adowa 123, Warszawa",
                                    "date_of_birth": "1990-01-15"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "User not found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/by_address": {
            "get": {
                "summary": "Get users by address",
                "parameters": [
                    {
                        "name": "address",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Address to search for (can be partial)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users matching the address",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "email": "user1@example.com",
                                        "phone_number": "+48123456789",
                                        "address": "ul. Przyk\u0142adowa 123, Warszawa",
                                        "date_of_birth": "1990-01-15"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Address parameter is required",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Address parameter is required"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}